Update code coverage m4
Previously if a user didn't have a particular (e.g. $FC) compiler
defined, because their code does not use that compiler and their
environment does not set it, the failure to compile code coverage with
"" as your compiler would cause configure to fail.  We now test for
compilers before trying to use them.

Previously we were seeing failures on OSX, where coverage flags must
*not* include an attempt to link with -lgcov.  We now test linker flags
with and without that library.